Migrated the jobs
