+++ package.json
@@ -50,11 +50,21 @@
     "@types/bent": "^7.3.2",
     "@types/express": "^4.17.11",
     "@types/git-diff": "^2.0.1",
+    "@types/jest": "^26.0.22",
     "@types/node-fetch": "^2.5.9",
     "@types/twit": "^2.2.28",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
+    "ts-jest": "^26.5.4",
     "typescript": "^4.2.3"
+  },
+  "jest": {
+    "roots": [
+      "<rootDir>/src"
+    ],
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    }
   },
   "nodemonConfig": {
     "verbose": true,
