+++ src/bot.ts
@@ -3,41 +3,40 @@
  * This code is licensed under MIT license (see LICENSE for details)
  */
 
+import { Client, Collection } from "discord.js";
 import { config } from "dotenv";
+import * as colors from "../colors.json";
+import * as reactions from "../reactions.json";
+import { CommandHandler } from "./handler/commands";
+import { MongoConnector } from "./handler/database";
+import { EventHandler } from "./handler/events";
+import { JobHandler } from "./handler/jobs";
+import { LogHandler } from "./handler/log";
+import { server } from "./server/server";
+import { BotType } from "./types/bot";
 config();
 
-import { server } from "./server/server";
 server.listen(process.env.PORT || 80, () => console.log(`💻 Webserver gestartet!`));
 
-import { BotType } from "./types/bot";
-
-import { Collection, Client } from "discord.js";
 const bot: BotType = new Client({ partials: ["MESSAGE", "CHANNEL", "REACTION"] });
 
 bot.config = process.env;
 bot.blacklist = new Collection();
 
-import * as colors from "../colors.json";
 bot.colors = colors;
 
-import * as reactions from "../reactions.json";
 bot.reactions = reactions;
 
-import { LogHandler } from "./handler/log";
 bot.logger = new LogHandler(bot);
 
-import { MongoConnector } from "./handler/database";
 const database = new MongoConnector();
 bot.MongoConnector = database;
 bot.db = database.connection;
 
-import { CommandHandler } from "./handler/commands";
 bot.commands = new CommandHandler(bot);
 
-import { JobHandler } from "./handler/jobs";
 bot.jobs = new JobHandler(bot);
 
-import { EventHandler } from "./handler/events";
 bot.events = new EventHandler(bot);
 bot.events.loadEvents();
