+++ tsconfig.json
@@ -70,5 +70,4 @@
     "forceConsistentCasingInFileNames": true,        /* Disallow inconsistently-cased references to the same file. */
     },
     "compileOnSave": true,
-
 }
